직원 유형을 정규직으로 변경하는 트랜잭션 구현. 테스트 코드 정상 작동. 에러 없음

+++ SalariedClassification.cpp
@@ -1,11 +1,11 @@
 #include "SalariedClassification.h"
 
-SalariedClassification::SalariedClassification(double salary) :
+SalariedClassification::SalariedClassification(float salary) :
 	itsSalary(salary)
 {
 }
 
-double const SalariedClassification::GetSalary()
+float const SalariedClassification::GetSalary()
 {
 	return itsSalary;
 }
